requirements.txt
APScheduler==3.11.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
packaging==25.0
premailer==3.10.0
pymongo==4.15.3
python-dotenv==1.1.1
requests==2.32.5
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yagmail==0.15.293
